fastApiNEXING/requirements.txt
fastapi
uvicorn
pandas
requests
pillow
ultralytics
openpyxl
io
